Update code to take in keyword data file
Created new scripts get_keyword_json.py to get all keyword dat in a JSON file.
This script may not be required in the future. It is being used now because MAG
embeddings are being generated on an Azure server that does not have access to the
Forward databse. To get around this, the keyword data is extracted into a file and
moved to the Azure server so the server doesn't have to rely on a Forward database.